refactor: adjusted query port path
