+++ src/main/java/pauloh/main/core/usecase/LoanService.java
@@ -10,9 +10,9 @@ import pauloh.main.adapter.input.dto.loan.LoanedBooksWithUsersReq;
 import pauloh.main.core.domain.model.Book;
 import pauloh.main.core.domain.model.Loan;
 import pauloh.main.port.input.LoanInputPort;
-import pauloh.main.port.input.LoanQueryPort;
 import pauloh.main.port.output.BookOutputPort;
 import pauloh.main.port.output.LoanOutputPort;
+import pauloh.main.port.output.LoanQueryPort;
 
 public class LoanService implements LoanInputPort {
   private final LoanOutputPort loanOutputPort;
