+++ src/main/java/pauloh/main/port/output/LoanQueryPort.java
@@ -1,4 +1,4 @@
-package pauloh.main.port.input;
+package pauloh.main.port.output;
 
 import java.util.List;
